Move AVR backlight to own file, add borrowed ARM implementation
