+++ keyboards/handwired/onekey/bluepill/config.h
@@ -21,3 +21,8 @@
 #define MATRIX_COL_PINS { B0 }
 #define MATRIX_ROW_PINS { A7 }
 #define UNUSED_PINS
+
+#define BACKLIGHT_PIN           A0
+#define BACKLIGHT_PWM_DRIVER    PWMD2
+#define BACKLIGHT_PWM_CHANNEL   1
+#define BACKLIGHT_BREATHING
