+++ keyboards/handwired/onekey/keymaps/backlight/rules.mk
@@ -1,0 +1,3 @@
+BACKLIGHT_ENABLE = yes
+CONSOLE_ENABLE = yes
+TAP_DANCE_ENABLE = yes
