+++ keyboards/handwired/onekey/stm32f0_discovery/rules.mk
@@ -1,34 +1,23 @@
-# GENERIC STM32F103C8T6 board - stm32duino bootloader
-BOARD = GENERIC_STM32_F103
-
-OPT_DEFS = -DCORTEX_VTOR_INIT=0x2000
-MCU_LDSCRIPT = STM32F103x8_stm32duino_bootloader
-
-DFU_ARGS = -d 1eaf:0003 -a2 -R
-DFU_SUFFIX_ARGS ?= -v 1eaf -p 0003
-
-# OPT_DEFS =
-# MCU_LDSCRIPT = STM32F103x8
-# BOARD = GENERIC_STM32_F103
-
 ## chip/board settings
 # the next two should match the directories in
 #  <chibios>/os/hal/ports/$(MCU_FAMILY)/$(MCU_SERIES)
 MCU_FAMILY = STM32
-MCU_SERIES = STM32F1xx
+MCU_SERIES = STM32F0xx
 # linker script to use
 # it should exist either in <chibios>/os/common/ports/ARMCMx/compilers/GCC/ld/
 #  or <this_dir>/ld/
+MCU_LDSCRIPT = STM32F072xB
 # startup code to use
 # is should exist in <chibios>/os/common/ports/ARMCMx/compilers/GCC/mk/
-MCU_STARTUP = stm32f1xx
+MCU_STARTUP = stm32f0xx
 # it should exist either in <chibios>/os/hal/boards/
 #  or <this_dir>/boards
+BOARD = ST_STM32F072B_DISCOVERY
 # Cortex version
 # Teensy LC is cortex-m0; Teensy 3.x are cortex-m4
-MCU  = cortex-m3
+MCU  = cortex-m0
 # ARM version, CORTEX-M0/M1 are 6, CORTEX-M3/M4/M7 are 7
-ARMV = 7
+ARMV = 6
 # If you want to be able to jump to bootloader from firmware on STM32 MCUs,
 # set the correct BOOTLOADER_ADDRESS. Either set it here, or define it in
 # ./bootloader_defs.h or in ./boards/<FOO>/bootloader_defs.h (if you have
@@ -39,3 +28,7 @@ ARMV = 7
 # This also requires a patch to chibios:
 #   <tmk_dir>/tmk_core/tool/chibios/ch-bootloader-jump.patch
 #STM32_BOOTLOADER_ADDRESS = 0x1FFFC800
+
+# Options to pass to dfu-util when flashing
+DFU_ARGS = -d 0483:df11 -a 0 -s 0x08000000:leave
+DFU_SUFFIX_ARGS = -p df11 -v 0483
